% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanuplists.R
\name{cleanuplists}
\alias{cleanuplists}
\title{Consolidate counts for identical capture histories}
\usage{
cleanuplists(zmse, includezerocounts = F)
}
\arguments{
\item{includezerocounts}{If F then remove zero rows.  If T then include all possible capture histories including those with zero count}
}
\description{
Consolidate counts for identical capture histories
}
