% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colmerge.R
\name{colmerge}
\alias{colmerge}
\title{Preprocess incidence matrix}
\usage{
colmerge(zmse, i = 0, j, rowclean = T, removezerocounts = T)
}
\arguments{
\item{zmse}{An incidence matrix with k+1 columns, where the first k columns are 1/0 giving the pattern of lists in that row, and the final column is the count}

\item{i}{If i is a positive number, merge lists i and j.  If i is a vector of positive numbers, merge all the lists in i. If i is a negative number or a vector of negative numbers, remove the corresponding lists.}

\item{j}{If i is a positive number, merge lists i and j. Otherwise j is ignored.  Note that to merge separate combinations of lists, or to both merge and omit, it is necessary to call the routine separately for each combination, noting that previous calls will alter the numbering of the lists}

\item{rowclean}{If rowclean=T then remove all redundant rows (either by combining rows with the same capture pattern or by removing rows with zero counts)}

\item{removezerocounts}{If this parameter is T then all rows with zero counts will be removed.  If F then return the full incidence matrix including combinations where the count is zero.}
}
\description{
This function preprocesses an incidence matrix of the form used in Rcapture.
 It allows you to merge two or more capture lists into one, or to omit one or more lists altogether.
 In addition it can produce a cleaned-up list where rows corresponding to a particular
     combination of lists are consolidated into a single row, and can produce a full dataset where the combinations with zero count
      are also included.
}
