% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madyorfunctions.R
\name{createarraydata}
\alias{createarraydata}
\title{Put data into correct format for the Madigan and York method}
\usage{
createarraydata(zdat)
}
\description{
Put data into correct format for the Madigan and York method
}
