% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{Kosovo}
\alias{Kosovo}
\title{Kosovo data}
\format{An object of class \code{data.frame} with 15 rows and 5 columns.}
\usage{
Kosovo
}
\description{
Data on 4400 observed killings in the Kosovo war between 20 March and 22 June 1999
}
\details{
These data (processed from \code{\link[LCMCR]{kosovo_aggregate}}) give the numbers of cases on each possible combination of four lists. 
The lists are labelled as follows: EXH = exhumations; ABA = American Bar Association Central and East European Law Initiative; OSCE = Organization for Security and Cooperation in Europe; HRW = Human Rights Watch.
All 15 combinations have a nonzero count.
}
\references{
Ball, P., W. Betts, F. Scheuren, J. Dudukovich, and J. Asher (2002). Killings and Refugee Flow in Kosovo March-June 1999. American Association for the Advancement of Science. A Report to the International Criminal Tribunal for the Former Yugoslavia.
}
\keyword{datasets}
