% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_AIC_stepwise_table1}
\alias{make_AIC_stepwise_table1}
\title{Script to return tables for the UK data}
\usage{
make_AIC_stepwise_table1()
}
\description{
Script to return tables for the UK data
}
