% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICfunctions.R
\name{make_AIC_stepwise_table2}
\alias{make_AIC_stepwise_table2}
\title{Script to return tables for the non-UK data sets}
\usage{
make_AIC_stepwise_table2()
}
\description{
Script to return tables for the non-UK data sets
}
