% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_allmodels_plots_script}
\alias{make_allmodels_plots_script}
\title{Plots of BIC against estimates for all pairwise models}
\usage{
make_allmodels_plots_script()
}
\description{
Produce all the plots in Section 3.3 of Silverman (2018)
}
