% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMCRfunctions.R
\name{make_LCMCR_table}
\alias{make_LCMCR_table}
\title{A function to make individual Dirichlet entries}
\usage{
make_LCMCR_table(zdat = UKdat, thin = 100, seed = 12345, burnin = 10000,
  ...)
}
\description{
A function to make individual Dirichlet entries
}
