% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_LCMCR_table_script}
\alias{make_LCMCR_table_script}
\title{A script to make the whole table}
\usage{
make_LCMCR_table_script(burnin0 = 1e+05)
}
\description{
A script to make the whole table
}
