% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madyorfunctions.R
\name{make_madyor_table}
\alias{make_madyor_table}
\title{Graphical models method}
\usage{
make_madyor_table(zdat, maxpop = NA, plotpost = F, returnmaxmodprob = T,
  ...)
}
\description{
Calling routine for the methods in the DGA package implementing
  the approach of Madigan and York (1997)
}
\references{
D. Madigan and J. C. York (1997). Bayesian methods for estimation of the size of a closed population. Biometrika 84, 19-31.
}
