% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_madyor_table_script}
\alias{make_madyor_table_script}
\title{Graphical Models script}
\usage{
make_madyor_table_script()
}
\description{
Script to make all the tables and figures in the section (5.1) 
 in Silverman (2018)
  on graphical models using the method of Madigan and York (1997)
}
\references{
D. Madigan and J. C. York (1997). Bayesian methods for estimation of the size of a closed population. Biometrika 84, 19-31.
}
