% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_MCMC_table}
\alias{make_MCMC_table}
\title{Table for Bayesian approach of Silverman (2018)}
\usage{
make_MCMC_table(zdata = UKdat, sigthreshseq = c(0, 2, 5), ...)
}
\description{
Find quantiles of the posterior distribution fo various 
 parameters of the prior and a specified range of values for
the thresholding step.
}
