% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_MCMCeffects_table_script}
\alias{make_MCMCeffects_table_script}
\title{Script for finding all the effects that survive thresholding for the various datasets}
\usage{
make_MCMCeffects_table_script(priorprec = 1)
}
\description{
Script for finding all the effects that survive thresholding for the various datasets
}
