% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_MCMCfit_tables_script}
\alias{make_MCMCfit_tables_script}
\title{Print out all the tables for the method of Silverman (2018)}
\usage{
make_MCMCfit_tables_script()
}
\description{
This gives all the tables for the Bayesian approach set out in Section 4 of Silverman (2018)
}
