% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCtables.R
\name{MCMCeffects}
\alias{MCMCeffects}
\title{Find effects that survive the threshold}
\usage{
MCMCeffects(zdata = UKdat, priorprec = 1, ...)
}
\description{
In the Bayesian thresholding method of Silverman (2018), find the effects
that survive thresholding at various threshold levels for given \code{priorprec}
}
