% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergelists}
\alias{mergelists}
\title{Merge specified lists into a single list}
\usage{
mergelists(zmse, i, j = NULL, k = NULL, l = NULL, rowclean = T,
  includezerocounts = F)
}
\arguments{
\item{i, j, k, l}{The lists to be merged.  Only i and j have to be given.}

\item{rowclean}{If rowclean=T then remove all redundant rows (either by combining rows with the
same capture pattern after the lists have been merged or by removing rows with zero counts)}
}
\description{
In order to merge distinct sets of lists, the routine has to be called for each set, noting that the column numbers
 will change after previous calls
}
