% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEfit.R
\name{MSEfit}
\alias{MSEfit}
\title{Fit MSE model using the Rcapture routine closedpCI.t}
\usage{
MSEfit(zdat = UKdat, mainonly = F, pthresh = 0.01)
}
\arguments{
\item{zdat}{Input data frame.  If there are k lists the first k columns are 1/0 to denote presence or absence on the relevant list; the last column is the count}

\item{mainonly}{If T then the stepwise method of BHS is used}

\item{pthresh}{The significance level that a new parameter has to reach to be included}
}
\description{
Implements the method of Bales, Hesketh and Silverman
}
