% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MSEretconf}
\alias{MSEretconf}
\title{Postprocess results from MSEfit}
\usage{
MSEretconf(zdat, zfit, mX, retfit)
}
\arguments{
\item{zdat}{Original data set}

\item{zfit}{The full fit as returned by \code{\link[Rcapture]{closedpCI.t}}}

\item{mX}{The particular model being fitted}

\item{retfit}{If T return the entire fit; if F return confidence intervals and the point estimate}
}
\description{
choose whether to return confidence levels or the full fit
}
