% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{Ned}
\alias{Ned}
\title{Netherlands data}
\format{An object of class \code{data.frame} with 24 rows and 7 columns.}
\usage{
Ned
}
\description{
Identified victims in The Netherlands, 2010-15
}
\details{
This is a table corresponding to six lists of victims in The Netherlands. The data are considered in six lists, labelled as follows: P = National Police; K = Border Police;
I = Inspectorate SZW (Ministry of Social Aairs and Employment); 
R = regional coordinators; O = residential treatment centers and shelters; Z = others
(for example, ambulatory care centers, organizations providing legal services,
Immigration and Naturalization Service). Each of the first six columns in the data frame
 corresponds to one of these lists.
  Each of the rows of the data frame corresponds to a possible combination of lists, with value 1 in the relevant
 column if the list is in that particular combination. 
 The last column of the data frame states the number of cases observed in that particular combination of lists. 
 Combinations of lists for which zero cases are observed are omitted.
}
\references{
J. J. van Dijk, M. Cruy, P. G. M. van der Heijden, and S. L. J. Kragten-Heerdink (2017). Monitoring Target 16.2 of the United Nations Sustainable Development Goals; a multiple systems estimation of the numbers of presumed human tracking victims in the Netherlands in 2010-2015 by year, age, gender, form of exploitation and nationality. Research Brief, available at \url{https://tinyurl.com/y9mpkach}.
}
\keyword{datasets}
