% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{Ned_5}
\alias{Ned_5}
\title{Netherlands data five list version}
\format{An object of class \code{data.frame} with 17 rows and 6 columns.}
\usage{
Ned_5
}
\description{
Netherlands data consolidated into five lists
}
\details{
This reduces the Netherlands data \code{\link{Ned}} into five lists, constructed by combining
 the two smallest lists I and O into a single list.
}
\keyword{datasets}
