% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{NewOrl}
\alias{NewOrl}
\title{New Orleans data}
\format{An object of class \code{data.frame} with 19 rows and 9 columns.}
\usage{
NewOrl
}
\description{
Victims related to modern slavery and human trafficking in New Orleans
}
\details{
These data are collected into 8 lists.  For reasons of confidentiality the labels of the lists are anonymised.
}
\references{
K. Bales, L. Murphy and B. W. Silverman (2018). How many trafficked and enslaved people are there in New Orleans? Unpublished report.
}
\keyword{datasets}
