% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{NewOrl_5}
\alias{NewOrl_5}
\title{New Orleans data five list version}
\format{An object of class \code{data.frame} with 14 rows and 6 columns.}
\usage{
NewOrl_5
}
\description{
New Orleans data consolidated into five lists
}
\details{
This reduces the New Orleans data \code{\link{NewOrl}} into five lists, constructed by combining
 the four smallest lists B, E, F and G into a single list.
}
\keyword{datasets}
