% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omitlists.R
\name{omitlists}
\alias{omitlists}
\title{Omit one or more of the lists in an incidence matrix}
\usage{
omitlists(zmse, i, j = NULL, k = NULL, l = NULL, rowclean = T,
  includezerocounts = F)
}
\arguments{
\item{i, j, k, l}{The lists to be omitted.  Only i has to be given.}

\item{rowclean}{If rowclean=T then combine rows with the same capture pattern}

\item{includezerocounts}{If F remove all rows with zero counts.  If T then include all possible observed capture combinations.}
}
\description{
Construct the incidence matrix obtained by omitting one or more of the original lists.  This is useful for diagnostic purposes
  and robustness testing, and also if some of the lists are less reliable.
}
