% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeemptyoverlaps.R
\name{removeemptyoverlaps}
\alias{removeemptyoverlaps}
\title{Remove pairs of lists with zero overlap}
\usage{
removeemptyoverlaps(zdata)
}
\value{
data The incidence matrix omitting all capture histories which contain a pair with zero overlap

removed  A matrix whose columns give the omitted pairs
}
\description{
All possible capture histories are included in the output, except for those which contain a pair of lists for which
 there is no overlap at all, whether or not in combination with other lists.
 So for a pair to be omitted, there must be zero counts in every capture history including both of those lists.
}
