% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{UKdat}
\alias{UKdat}
\title{UK data}
\format{An object of class \code{data.frame} with 25 rows and 7 columns.}
\usage{
UKdat
}
\description{
Data from the UK 2013 strategic assessment
}
\details{
This is a table of six lists used in the research \href{https://tinyurl.com/ydegfjaw}{published by the Home Office} as part of the strategy leading
 to the Modern Slavery Act 2015. The data are considered in six lists, labelled as follows:  LA--Local authorities; NG--Non-government organisations;
PF--Police forces; GO--Government organisations; GP--General public; NCA--National Crime Agency.  Each of the first six columns in the data frame
 corresponds to one of these lists.
  Each of the rows of the data frame corresponds to a possible combination of lists, with value 1 in the relevant
 column if the list is in that particular combination. 
 The last column of the data frame states the number of cases observed in that particular combination of lists. 
 Combinations of lists for which zero cases are observed are omitted.
}
\references{
\url{https://www.gov.uk/government/publications/modern-slavery-an-application-of-multiple-systems-estimation}
}
\keyword{datasets}
