% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{UKdat_4}
\alias{UKdat_4}
\title{UK data four list version}
\format{An object of class \code{data.frame} with 14 rows and 5 columns.}
\usage{
UKdat_4
}
\description{
UK data consolidated into four lists
}
\details{
This reduces the UK data \code{\link{UKdat}} into four lists, constructed by combining
 the PF and NCA lists into a single PFNCA list and by omitting the GP list
}
\keyword{datasets}
