% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetdocumentation.R
\docType{data}
\name{UKdat_5}
\alias{UKdat_5}
\title{UK data five list version}
\format{An object of class \code{data.frame} with 18 rows and 6 columns.}
\usage{
UKdat_5
}
\description{
UK data consolidated into five lists
}
\details{
This reduces the UK data \code{\link{UKdat}} into five lists, constructed by combining
 the PF and NCA lists into a single PFNCA list
}
\keyword{datasets}
