% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKdat.R
\docType{data}
\name{UKdat}
\alias{UKdat}
\title{UK NRM data 2013}
\format{An object of class \code{data.frame} with 25 rows and 7 columns.}
\usage{
UKdat
}
\description{
This is the United Kingdom National Referral Mechanism data as used by Bales, Hesketh and Silverman (2014)
}
\keyword{datasets}
